Move tests to packages
