+++ src/functionalTest/java/uk/gov/hmcts/reform/feature/api/DeleteFeatureToggleTest.java
@@ -1,9 +1,10 @@
-package uk.gov.hmcts.reform.feature;
+package uk.gov.hmcts.reform.feature.api;
 
 import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
 import org.junit.Test;
 import org.springframework.http.HttpStatus;
+import uk.gov.hmcts.reform.feature.BaseTest;
 
 import java.io.IOException;
 import java.util.UUID;
