+++ src/functionalTest/java/uk/gov/hmcts/reform/feature/api/FeatureStoreTest.java
@@ -1,7 +1,8 @@
-package uk.gov.hmcts.reform.feature;
+package uk.gov.hmcts.reform.feature.api;
 
 import io.restassured.path.json.JsonPath;
 import org.junit.Test;
+import uk.gov.hmcts.reform.feature.BaseTest;
 
 import java.io.IOException;
 import java.util.UUID;
